# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Portal Academico API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /teachers:
    get:
      operationId: listTeachers
      tags: [teachers]
      summary: List all teachers
      responses:
        "200":
          description: List of teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Create a teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeacherRequest"
      responses:
        "201":
          description: Created teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /teachers/{id}:
    put:
      operationId: updateTeacher
      tags: [teachers]
      summary: Update teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeacherRequest"
      responses:
        "200":
          description: Updated teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    delete:
      operationId: deleteTeacher
      tags: [teachers]
      summary: Delete teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [administrador, coordinador, administrativo]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - fullName
        - email
        - role
        - isActive
        - createdAt
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [administrador, coordinador, administrativo]
      required:
        - username
        - fullName
        - email
        - password
        - role
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [administrador, coordinador, administrativo]
        isActive:
          type: boolean
    Teacher:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        specialty:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - fullName
        - email
        - specialty
        - isActive
    CreateTeacherRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        specialty:
          type: string
        phone:
          type: string
      required:
        - fullName
        - email
        - specialty
...